Fix a bug with reexport-all
The prior code wasn't extracting the member part of the token name
when reexporting symbols, meaning the member was fully qualified.  For
example, when reexporting a class "C" from another module "other",
the resulting name was "other:C".  This is correct for the export node's
token -- so it can be resolved correctly to its source -- however,
clearly the identifier for the export should not contain the module
part.  This change strips it in the appropriate place.