Add a visitation API
This change introduces a new visitation API to the new MuIL AST.
The ast.Walk API takes an ast.Visitor implementation and walks the
tree in depth-first order, invoking the visitor along the way.

The visitor gets to choose whether to continue visitation (by returning
a non-nil visitor object), or to stop it (by returning nil).  The
visitation will proceed with that returned visitor, so that a visitor
can "swap out" the visitor used for child nodes if needed.

At the end, the PostVisit function is called, for any clean up logic.

Finally, the ast.Inspector type is available as a simple way of consing
up visitors simply using a function that returns a bool indicating
whether visitation should continue.