Add support for loading locations
This adds support for LoadLocationExpression, which will be key to
getting virtually everything else working, including class and module
member accesses, local variable gets/sets, function calls, and more.

In particular, an extra level of indirection is added to the globals,
locals, and properties maps.  This new thing is called a *Reference,
and it is simply a mutable slot that contains an *Object.  Instead
of storing and retrieving objects directly, this extra level of
indirection enables us to create *Object instances that themselves
simply refer to other objects (for subsequent loads and stores).

For now, we do not permit overwriting of method properties.  This
is a future work item (see marapongo/mu#56).