Encapsulate Unwind's innards
This change uses an enum rather than set of bools for unwind's inner
state, makes its fields private, and adds accessor methods.  This makes
it easier to protect invariants and wonky states like being a break and
a throw simultaneously.  Now that Unwind is part of the public API (vs
being private as it used to), this is really a requirement (and was
obviously a good idea regardless).  Thanks to Eric for CR feedback.