Emit good tokens
This overhauls how MuJS creates and emits tokens, so that they are
properly fully qualified.  Before this change, we frequently emitted
"naked" identifier names that, though the compiler had already resolved
and bound them, looked to the MuIL toolchain as though they hadn't been.

This required rearranging the "symbols" (now "tokens") layer to look more
like the types in the MuIL toolchain.

More impactfully, however, it means we need to load MuIL packages from
dependencies as part of creating the fully resolved tokens.  This injects
I/O into places it didn't exist before, so a large part of this change is
simply playing the async/await whack-a-mole game.

This is one big step along the path of fixing marapongo/mu#46.  It
still isn't 100% because we need to emit array/property symbols correctly,
and also I suspect as we do more multi-module testing, a few other things
will crop up.  But we're almost there.