Invoke the OnVariableAssign interpreter hook
This change actually invokes the OnVariableAssign interpreter hook
at the right places.  (Renamed from OnAssignProperty, as it will now
handle all variable assignments, and not just properties.)  This
requires tracking a bit more information about l-values so that we
can accurately convey the target object and symbol associated with
the assignment (resulting in the new "location" struct type).