Generate correct array type tokens
This change adds support for generating array type tokens from MuJS.
Previously, we'd try to load up the Array type's associate MuPackage,
which is TypeScript's standard library; and then we'd error out
because, of course, it doesn't have one.  Now we recognize references
to it (and ES6-style Map<K, V>) to generate the correct tokens.

This includes a bunch of new array tests.