Add map tests
This adds map tests for ES6-style maps encoded in MuPack/MuIL.

It also fixes one simple bug, where we should not populate the
compiler options with TypeScript defaults *unless* they are (a)
missing from the arguments and (b) not discovered from a tsconfig
file.  Otherwise, we will rewrite overrides with the defaults.