Save MuJS output to Mupack.json by default
This change stops printing the AST to output after compilation.  It's
a little confusing and, especially as Mu test programs are getting larger,
staring at 1,000s of lines of JSON whiz by is losing it's fun.

By default, the compiler saves output into Mupack.json.  This can be
overridden with the --out switch (-o for short).  Both JSON and YAML files
are supported (as are unrecognized extensions, which get JSON by default).
Another switch --format (-f for short) is supported to explicitly specify
the output format.  A special -out of "-" means "print to stdoutput", so
for instance, running `mujs -o=- -f=yaml` will print YAML to the console.