Implement multi-pass module binding
The prior module binding logic was naive.  It turns out that, because of inter-module
references within the same package, the order of binding certain modules and its members
within a package matters.  This leads to a more Java/C#/JavaScript/etc. multi-pass model,
versus a classical single-pass C/C++ model (which requires forward declarations).

To do this, we bind things in this order within a package:

* First, we add an entry for each module.  It is largely empty, but at least names resolve.

* Second, we bind all imports.  Since all modules are known, all inter-module tokens resolve.

* Next, we can bind class symbols.  Note that we must do this prior to module properties,
  methods, and exports, since they might be referenced and need to exist first.  As before,
  we do not yet bind function bodies, since those can reference anything.

* Now we bind all module "members", meaning any module scoped properties and methods.  Again,
  we do not bind function bodies just yet.  This is just a symbolic binding step.

* Next, we can finally bind exports.  Because exports may reference any members, all of the
  above must have been done first.  But all possibly exported symbols are now resolved.

* Finally, we can bind function bodies.