Bind class member definitions in a separate pass
This change further rearranges the phasing of binding to account for
the fact that class definitions may freely reference exports, which won't
be known until after binding class names.  Therefore, we bind class names
first and then the definitions (function signatures and variables).

This brings the AWS MuPackage's number of verification errors down to 84
from 136.