Use tsconfig.json's outDir by default
There are a variety of ways to specify a MuJS output location, including
the command line --out (-o for short) flag.  If that isn't present, we simply
dump the Mupack.json file in the current working directory.  However, in order
to prepare Mupackage files for packaging, distribution, and reuse, we actually
want to store the Mupack.json file alongside all the other package assets (like
TypeScript header files).  So it makes sense to recognize the outDir, if present,
in the tsconfig.json file, and use it as the default if not otherwise specified.