Save TypeScript definition files
To prepare a MuPackage for consumption by another MuPackage, we need
to save the definition files output by the TypeScript compiler.  In
theory, we could regenerate these definitions from the MuPackage itself;
in fact, when we tackle cross-language, this will be necessary.  For
now, however, it's much simpler to simply save the definition outputs
alongside the Mupack.json package metadata.

Note that we do not save "code" outputs.  The MuIL format encodes all
computations and so the only thing we need for dependent packages is
the header files to compile against; all "link" and runtime dependencies
are satisfied by the MuIL format and Mu toolchain.