Implement basic default modules and entrypoints
Per marapongo/mu#57, this change marks the "index" module as the default.
We don't yet discover and crack open `package.json` files which may override
this by specifying a "main" function.

This change also generates entrypoint functions for all MuJS modules.  To
emulate the way Node.js works -- in that every module is still a valid script
target -- we are generating entrypoints for everything; this is in contrast to
only generating them for "blueprint" modules, or somehow marking `main`
somehow.  See the notes in marapongo/mu#57 for additional thoughts on this.